fix(pci-public-ip): fix code review proposition
ref: DTCORE-2010
